// shared/api/generated.ts
export interface SkillDto {
  id: number;
  title: string;
}

export interface ProjectDto {
  deploy: string;
  description: string;
  github: string;
  id: number;
  image: string;
  labels: string;
  skills: SkillDto[];
  title: string;
}